+++ CHANGELOG.md
@@ -35,6 +35,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fix cassette playing weapons not having a menu for storing cassette
 - Fix weapons not being enchantable
 - Add back song description for cassette items
+- Fix hover area for some tooltip in the echeladder screen
 
 ### Removed